+++ src/Symfony/Component/Cache/Adapter/SimpleCacheAdapter.php
@@ -11,69 +11,11 @@
 
 namespace Symfony\Component\Cache\Adapter;
 
-use Psr\SimpleCache\CacheInterface;
-use Symfony\Component\Cache\PruneableInterface;
-use Symfony\Component\Cache\ResettableInterface;
-use Symfony\Component\Cache\Traits\ProxyTrait;
+@trigger_error(sprintf('The "%s" class is @deprecated since Symfony 4.3, use "Psr16Adapter" instead.', SimpleCacheAdapter::class), E_USER_DEPRECATED);
 
 /**
- * @author Nicolas Grekas <p@tchwork.com>
+ * @deprecated since Symfony 4.3, use Psr16Adapter instead.
  */
-class SimpleCacheAdapter extends AbstractAdapter implements PruneableInterface, ResettableInterface
+class SimpleCacheAdapter extends Psr16Adapter
 {
-    use ProxyTrait;
-
-    private $miss;
-
-    public function __construct(CacheInterface $pool, string $namespace = '', int $defaultLifetime = 0)
-    {
-        parent::__construct($namespace, $defaultLifetime);
-
-        $this->pool = $pool;
-        $this->miss = new \stdClass();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function doFetch(array $ids)
-    {
-        foreach ($this->pool->getMultiple($ids, $this->miss) as $key => $value) {
-            if ($this->miss !== $value) {
-                yield $key => $value;
-            }
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function doHave($id)
-    {
-        return $this->pool->has($id);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function doClear($namespace)
-    {
-        return $this->pool->clear();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function doDelete(array $ids)
-    {
-        return $this->pool->deleteMultiple($ids);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function doSave(array $values, $lifetime)
-    {
-        return $this->pool->setMultiple($values, 0 === $lifetime ? null : $lifetime);
-    }
 }