+++ src/Symfony/Component/Cache/Tests/Simple/AbstractRedisCacheTest.php
@@ -13,6 +13,9 @@ namespace Symfony\Component\Cache\Tests\Simple;
 
 use Symfony\Component\Cache\Simple\RedisCache;
 
+/**
+ * @group legacy
+ */
 abstract class AbstractRedisCacheTest extends CacheTestCase
 {
     protected $skippedTests = [