+++ src/Symfony/Component/Cache/Tests/Simple/Psr6CacheWithAdapterTest.php
@@ -15,6 +15,7 @@ use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class Psr6CacheWithAdapterTest extends Psr6CacheTest
 {